Fix some narrowing conversions inside static initializers that prevent VisIt from building with C++11.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/branches/2.11RC@29352 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/VisWindow/Colleagues/avtImageColleague.C
@@ -560,7 +560,9 @@ avtImageColleague::GetOptions(AnnotationObject &annot)
 
     annot.SetPosition(actor->GetPosition());
 
-    double pos[] = {width, height, 0};
+    double pos[] = {static_cast<double>(width),
+                    static_cast<double>(height),
+                    0.};
     annot.SetPosition2(pos);
 
     annot.SetIntAttribute1(useOpacityColor ? 1 : 0);
